requirements.txt
pyyaml == 6.0.1
pynput == 1.7.6